Consolidate tests into mirrored files.
For example, tests for all code in "myfile.py" will be held in
"test_myfile.py".
